requirements.txt
torch>=2.1.0
transformers>=4.35.0
Pillow>=10.0.0
gradio>=3.44.0